refactor(query_api): extract oracle run/eval/status methods to oracle_runs.rs
Move run_oracle(_at/_with_tool/_from_scip), oracle_pre_spawn_snapshot,
oracle_eval_metrics, oracle_status, probe_oracle_tool, latest_oracle_run_*
into oracle_runs.rs. Named oracle_runs (not oracle) to avoid shadowing the
existing `use crate::index::oracle` referenced as bare oracle:: throughout.
mod.rs's oracle type imports (only the moved methods used them) dropped to
`use crate::index::oracle`. No behavior change.
