+++ crates/rag-rat-core/src/index/query_api/oracle_runs.rs
@@ -1,0 +1,171 @@
+use std::path::Path;
+
+use super::*;
+use crate::index::oracle::{
+    self, OracleEvalMetrics, OracleReport, OracleStatus, OracleTool, RecallCalls,
+};
+
+impl IndexDatabase {
+    pub fn run_oracle(
+        &self,
+        tool: OracleTool,
+        tool_version: &str,
+        scip_bytes: &[u8],
+        production_sha: Option<&std::collections::HashMap<String, String>>,
+        pre_spawn_sha: Option<&std::collections::HashMap<String, String>>,
+    ) -> anyhow::Result<OracleReport> {
+        self.run_oracle_at(tool, tool_version, scip_bytes, production_sha, pre_spawn_sha, now_ms())
+    }
+
+    /// As [`Self::run_oracle`], but records the run's `started_at` as `started_at_ms` — the moment
+    /// the caller began the run (its pre-spawn snapshot), not completion time. The tool-driven path
+    /// passes this so the auto-run staleness gate isn't wedged by a run that overlapped a watcher
+    /// reindex (#145).
+    pub fn run_oracle_at(
+        &self,
+        tool: OracleTool,
+        tool_version: &str,
+        scip_bytes: &[u8],
+        production_sha: Option<&std::collections::HashMap<String, String>>,
+        pre_spawn_sha: Option<&std::collections::HashMap<String, String>>,
+        started_at_ms: i64,
+    ) -> anyhow::Result<OracleReport> {
+        let Some(root) = self.storage.source_root() else {
+            anyhow::bail!(
+                "index has no source_root metadata; rebuild required for the oracle pass"
+            );
+        };
+        let root = root.to_path_buf();
+        oracle::run_oracle_at(
+            self.storage.connection(),
+            tool,
+            tool_version,
+            &self.active_commit_sha,
+            &self.active_worktree_id,
+            scip_bytes,
+            &root,
+            production_sha,
+            pre_spawn_sha,
+            started_at_ms,
+        )
+    }
+
+    /// The active checkout's indexed `(path -> files.sha256)` map — the pre-spawn snapshot the
+    /// CLI takes BEFORE spawning the oracle tool (and before acquiring the index write lock; this
+    /// is a cheap read-only query), so the join can reject any document the watcher reindexed
+    /// across the entire spawn → join window (#83).
+    pub fn oracle_pre_spawn_snapshot(
+        &self,
+    ) -> anyhow::Result<std::collections::HashMap<String, String>> {
+        oracle::pre_spawn_snapshot(
+            self.storage.connection(),
+            &self.active_commit_sha,
+            &self.active_worktree_id,
+        )
+    }
+
+    /// Heuristic-vs-oracle eval metrics (precision/recall/recovery) for a tool/version, diffing the
+    /// persisted `edge_oracle` rows against the `edges` heuristic. [`RecallCalls`] is the
+    /// `(covered_calls, oracle_only_calls)` pair reported by the most recent [`run_oracle`] — both
+    /// occurrence-counted over the call population, so recall compares like with like.
+    pub fn oracle_eval_metrics(
+        &self,
+        tool: OracleTool,
+        tool_version: &str,
+        recall_calls: RecallCalls,
+    ) -> anyhow::Result<OracleEvalMetrics> {
+        oracle::oracle_eval_metrics(
+            self.storage.connection(),
+            tool,
+            tool_version,
+            &self.active_commit_sha,
+            &self.active_worktree_id,
+            recall_calls,
+        )
+    }
+
+    /// Persisted oracle status (verdict counts + last run) for a tool/version, scoped to this
+    /// database's active `(commit_sha, worktree_id)` checkout.
+    pub fn oracle_status(
+        &self,
+        tool: OracleTool,
+        tool_version: &str,
+    ) -> anyhow::Result<OracleStatus> {
+        oracle::oracle_status(
+            self.storage.connection(),
+            tool,
+            tool_version,
+            &self.active_commit_sha,
+            &self.active_worktree_id,
+        )
+    }
+
+    /// `rag-rat oracle run [--tool <id>]` without a pre-built `--scip`: invoke the indexer to
+    /// produce a `.scip` (to a caller-owned temp path), then run the phase-1 join over it. A
+    /// missing or unrunnable tool returns [`oracle::OracleRunOutcome::Blocked`] with an install
+    /// hint (the CLI prints it and exits 0) — never an error. Records an `oracle_runs` row on
+    /// success.
+    pub fn run_oracle_with_tool(
+        &self,
+        tool: OracleTool,
+        scip_output: &Path,
+    ) -> anyhow::Result<oracle::OracleRunOutcome> {
+        let Some(root) = self.storage.source_root() else {
+            anyhow::bail!(
+                "index has no source_root metadata; rebuild required for the oracle pass"
+            );
+        };
+        let root = root.to_path_buf();
+        oracle::run_oracle_with_tool(
+            self.storage.connection(),
+            tool,
+            &root,
+            scip_output,
+            &self.active_commit_sha,
+            &self.active_worktree_id,
+        )
+    }
+
+    /// Run the oracle pass over a PRE-BUILT `.scip` for a tool, recording an `oracle_runs` row. The
+    /// `--scip <path>` consumption path of `oracle run`; deterministic (no subprocess), so it's the
+    /// tested end-to-end seam. `tool_version` labels the run (content-addressed staleness key).
+    pub fn run_oracle_from_scip(
+        &self,
+        tool: OracleTool,
+        tool_version: &str,
+        scip_bytes: &[u8],
+    ) -> anyhow::Result<oracle::OracleReport> {
+        // A pre-built `--scip` carries no production moment or spawn we control, so neither the
+        // scip-vs-disk nor the pre-spawn gate can arm — only the index-vs-disk gate applies.
+        self.run_oracle(tool, tool_version, scip_bytes, None, None)
+    }
+
+    /// Probe whether an oracle tool is installed, for `oracle status`. A `Blocked` probe is
+    /// informational (the tool isn't installed), never an error.
+    pub fn probe_oracle_tool(&self, tool: OracleTool) -> oracle::ToolAvailability {
+        oracle::probe_oracle_tool(tool)
+    }
+
+    /// The `tool_version` of the most recent oracle run for `tool` in this checkout, or `None` when
+    /// no run exists. The version `oracle status` reports verdict counts against.
+    pub fn latest_oracle_run_version(&self, tool: OracleTool) -> anyhow::Result<Option<String>> {
+        oracle::latest_run_tool_version(
+            self.storage.connection(),
+            tool,
+            &self.active_commit_sha,
+            &self.active_worktree_id,
+        )
+    }
+
+    /// The `started_at` (Unix-epoch ms) of the most recent oracle run for `tool` in this checkout,
+    /// or `None` when none exists — the staleness clock the background auto-fresh oracle (Phase
+    /// 5) compares against the index's `indexed_at_ms` to decide whether verdicts are stale.
+    pub fn latest_oracle_run_started_at(&self, tool: OracleTool) -> anyhow::Result<Option<i64>> {
+        oracle::latest_run_started_at(
+            self.storage.connection(),
+            tool,
+            &self.active_commit_sha,
+            &self.active_worktree_id,
+        )
+    }
+}
